Make some chenges after review, but not all

+++ StoreProject/Server/src/main/java/api/dao/GenericDao.java
@@ -6,7 +6,7 @@ public interface GenericDao<T> {
 
     void create(T entity);
 
-    T getById(int id);
+    T getById(Long id);
 
     List<T> getAll();
